+++ packages/@tailwindcss-upgrade/src/codemods/template/candidates.ts
@@ -1,4 +1,6 @@
 import { Scanner } from '@tailwindcss/oxide'
+import type { Candidate } from '../../../../tailwindcss/src/candidate'
+import type { DesignSystem } from '../../../../tailwindcss/src/design-system'
 
 export async function extractRawCandidates(
   content: string,
@@ -12,4 +14,22 @@ export async function extractRawCandidates(
     candidates.push({ rawCandidate, start, end: start + rawCandidate.length })
   }
   return candidates
+}
+
+// Create a basic stripped candidate without variants or important flag
+export function baseCandidate<T extends Candidate>(candidate: T) {
+  let base = structuredClone(candidate)
+
+  base.important = false
+  base.variants = []
+
+  return base
+}
+
+export function parseCandidate(designSystem: DesignSystem, input: string) {
+  return designSystem.parseCandidate(
+    designSystem.theme.prefix && !input.startsWith(`${designSystem.theme.prefix}:`)
+      ? `${designSystem.theme.prefix}:${input}`
+      : input,
+  )
 }